pyproject.toml
[project]
name = "nobus-cloud-assingments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-ninja>=1.5.3",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
]
